+++ src/Molecules/Input/Select/lib/wrappers/FormFieldOrFragment.tsx
@@ -88,6 +88,7 @@ export const FormFieldOrFragment = React.forwardRef<HTMLDivElement, any>(
       size,
       labelToolTip,
       labelTooltipPosition,
+      labelTooltipInModal,
       innerRef,
       ...props
     },
@@ -121,6 +122,7 @@ export const FormFieldOrFragment = React.forwardRef<HTMLDivElement, any>(
             fieldId={id}
             labelTooltip={labelToolTip}
             labelTooltipPosition={labelTooltipPosition}
+            labelTooltipInModal={labelTooltipInModal}
             {...props}
             {...(fullWidth ? { width: '100%' } : {})}
             ref={ref}