+++ src/Molecules/Input/Text/Text.tsx
@@ -222,6 +222,7 @@ const TextComponent = React.forwardRef<HTMLInputElement, Props>((props, ref) => 
           'label',
           'labelTooltip',
           'labelTooltipPosition',
+          'labelTooltipInModal',
           'className',
           'width',
           'disabled',