+++ src/Molecules/SelectionCard/SelectionCard.stories.tsx
@@ -2,9 +2,11 @@ import React, { useState } from 'react';
 import { Meta, Story } from '@storybook/react';
 import styled from 'styled-components';
 
-import { Flexbox, OldIcon, Typography } from '../..';
+import { Box, Flexbox, Icon, Typography } from '../..';
 import { SelectionCard } from './SelectionCard';
 import { SelectionCard as Props } from './SelectionCard.types';
+
+import image from './images/guidance_economic_independence.jpg';
 
 export default {
   title: 'Molecules / SelectionCard',
@@ -13,6 +15,11 @@ export default {
 
 const StyledFlexbox = styled(Flexbox)`
   width: 100%;
+`;
+
+const StyledBox = styled(Box)`
+  height: 100%;
+  width: 150px;
 `;
 
 const Template: Story<Props> = (args) => <SelectionCard {...args} />;
@@ -55,8 +62,26 @@ export const WithIcon = Template.bind({});
 WithIcon.args = {
   ...SelectionCardDefault,
   title: 'With Icon',
-  icon: <OldIcon.House size={8} />,
+  icon: <Icon.Book24 />,
 };
+
+export const withImage = () => {
+  const Component = () => {
+    return (
+      <StyledBox>
+        <SelectionCard
+          title="Controlled selection card"
+          text="This component is controlled"
+          imageUrl={image}
+          border
+        />
+      </StyledBox>
+    );
+  };
+  return <Component />;
+};
+
+withImage.storyName = 'With image';
 
 export const withValueControlledBehavior = () => {
   const Component = () => {